More cleanup and one important fix
Get rid of aria2 -- yet another tool I haven't used for years and only
used it once for something.

Installer behavior: for the love of all that is good and holy, deliver
me from my fat fingers accidentally pressing a key and WHOOPS, it skips
something because it expected a single character (y or Y) to confirm...
Wait for ENTER from now on.

Some other stuff